+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-### Added
+### Fixed
 
-- Add useIOResources hook
+- Mem leak caused by pruneCache holding refs to all caches
+
+### Removed
+
+- pruneCache
 
 ### Changed